+++ crates/client/src/client/types.rs
@@ -43,6 +43,7 @@ use crate::client::schema::{
     relayed_tx::RelayedTransactionStatus as SchemaRelayedTransactionStatus,
     tx::{
         OpaqueTransactionWithStatus,
+        StatusWithTransaction as SchemaStatusWithTx,
         TransactionStatus as SchemaTxStatus,
     },
     ConversionError,
@@ -95,13 +96,16 @@ pub struct TransactionResponse {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub struct StatusWithTransactionResponse {
+    pub status: StatusWithTransaction,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub enum TransactionStatus {
     Submitted {
         submitted_at: Tai64,
     },
     Success {
-        #[cfg(feature = "test-helpers")]
-        transaction: Transaction,
         block_height: BlockHeight,
         time: Tai64,
         program_state: Option<ProgramState>,
@@ -113,8 +117,6 @@ pub enum TransactionStatus {
         reason: String,
     },
     Failure {
-        #[cfg(feature = "test-helpers")]
-        transaction: Transaction,
         block_height: BlockHeight,
         time: Tai64,
         reason: String,
@@ -134,8 +136,6 @@ impl TryFrom<SchemaTxStatus> for TransactionStatus {
                 submitted_at: s.time.0,
             },
             SchemaTxStatus::SuccessStatus(s) => TransactionStatus::Success {
-                #[cfg(feature = "test-helpers")]
-                transaction: s.transaction.try_into()?,
                 block_height: s.block_height.into(),
                 time: s.time.0,
                 program_state: s.program_state.map(TryInto::try_into).transpose()?,
@@ -148,8 +148,6 @@ impl TryFrom<SchemaTxStatus> for TransactionStatus {
                 total_fee: s.total_fee.0,
             },
             SchemaTxStatus::FailureStatus(s) => TransactionStatus::Failure {
-                #[cfg(feature = "test-helpers")]
-                transaction: s.transaction.try_into()?,
                 block_height: s.block_height.into(),
                 time: s.time.0,
                 reason: s.reason,
@@ -166,6 +164,80 @@ impl TryFrom<SchemaTxStatus> for TransactionStatus {
                 TransactionStatus::SqueezedOut { reason: s.reason }
             }
             SchemaTxStatus::Unknown => {
+                return Err(Self::Error::UnknownVariant("SchemaTxStatus"))
+            }
+        })
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub enum StatusWithTransaction {
+    Submitted {
+        submitted_at: Tai64,
+    },
+    Success {
+        transaction: Transaction,
+        block_height: BlockHeight,
+        time: Tai64,
+        program_state: Option<ProgramState>,
+        receipts: Vec<Receipt>,
+        total_gas: u64,
+        total_fee: u64,
+    },
+    SqueezedOut {
+        reason: String,
+    },
+    Failure {
+        transaction: Transaction,
+        block_height: BlockHeight,
+        time: Tai64,
+        reason: String,
+        program_state: Option<ProgramState>,
+        receipts: Vec<Receipt>,
+        total_gas: u64,
+        total_fee: u64,
+    },
+}
+
+impl TryFrom<SchemaStatusWithTx> for StatusWithTransaction {
+    type Error = ConversionError;
+
+    fn try_from(status: SchemaStatusWithTx) -> Result<Self, Self::Error> {
+        Ok(match status {
+            SchemaStatusWithTx::SubmittedStatus(s) => StatusWithTransaction::Submitted {
+                submitted_at: s.time.0,
+            },
+            SchemaStatusWithTx::SuccessStatus(s) => StatusWithTransaction::Success {
+                transaction: s.transaction.try_into()?,
+                block_height: s.block_height.into(),
+                time: s.time.0,
+                program_state: s.program_state.map(TryInto::try_into).transpose()?,
+                receipts: s
+                    .receipts
+                    .into_iter()
+                    .map(TryInto::try_into)
+                    .collect::<Result<Vec<_>, _>>()?,
+                total_gas: s.total_gas.0,
+                total_fee: s.total_fee.0,
+            },
+            SchemaStatusWithTx::FailureStatus(s) => StatusWithTransaction::Failure {
+                transaction: s.transaction.try_into()?,
+                block_height: s.block_height.into(),
+                time: s.time.0,
+                reason: s.reason,
+                program_state: s.program_state.map(TryInto::try_into).transpose()?,
+                receipts: s
+                    .receipts
+                    .into_iter()
+                    .map(TryInto::try_into)
+                    .collect::<Result<Vec<_>, _>>()?,
+                total_gas: s.total_gas.0,
+                total_fee: s.total_fee.0,
+            },
+            SchemaStatusWithTx::SqueezedOutStatus(s) => {
+                StatusWithTransaction::SqueezedOut { reason: s.reason }
+            }
+            SchemaStatusWithTx::Unknown => {
                 return Err(Self::Error::UnknownVariant("SchemaTxStatus"))
             }
         })